creation of database and database schema
